feat: add skill entity linking for project assistant
Teach the dashboard assistant to emit gram:skill/<id> links and resolve
them to the project skill detail page so newly created skills are
clickable in chat replies.
