+++ server/internal/assistants/source_adapter_test.go
@@ -104,6 +104,9 @@ func TestComposeInstructions_DashboardIncludesElementsPrompts(t *testing.T) {
 	require.Contains(t, instructions, "Only render ONE generative UI widget")
 	require.Contains(t, instructions, "BarChart")
 	require.Contains(t, instructions, "```ui code blocks")
+	require.Contains(t, instructions, "## Linking entities")
+	require.Contains(t, instructions, "gram:skill/<ID>")
+	require.Contains(t, instructions, "Skill.ID from the skill tools")
 }
 
 func TestDashboardAdapterDecodeTurnIncludesSelectedSkills(t *testing.T) {
